ml: introduce generic env trait
