+++ crates/ml/src/lib.rs
@@ -2,6 +2,7 @@ pub mod graph;
 pub mod nn;
 pub mod optim;
 pub mod recorder;
+pub mod env;
 pub mod rl;
 pub mod stick_balance;
 pub mod tape;
@@ -9,3 +10,4 @@ pub mod tensor;
 
 pub use tensor::Tensor;
 pub use stick_balance::StickBalanceEnv;
+pub use env::Env;
